requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.3
altair==4.2.2
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
asgiref==3.6.0
asteroid-filterbanks==0.4.0
async-timeout==4.0.2
attrs==22.2.0
audioread==3.0.0
backoff==2.2.1
backports.cached-property==1.0.2
bentoml==1.0.15
blinker==1.5
blis==0.7.9
boto3==1.26.74
botocore==1.29.74
Brotli==1.0.9
build==0.10.0
cachetools==5.3.0
catalogue==2.0.8
cattrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
circus==0.18.0
click==8.1.3
click-option-group==0.5.5
cloudpickle==2.2.1
cmaes==0.9.1
cmake==3.26.1
colorama==0.4.6
colorlog==6.7.0
confection==0.0.4
contextlib2==21.6.0
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
decorator==5.1.1
deepmerge==1.1.0
Deprecated==1.2.13
docker==5.0.3
docopt==0.6.2
einops==0.3.2
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.4.1/en_core_web_md-3.4.1-py3-none-any.whl
entrypoints==0.4
exceptiongroup==1.1.0
fastapi==0.95.0
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.9.0
fonttools==4.39.3
frozenlist==1.3.3
fs==2.4.16
fsspec==2023.4.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.2
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.56.0
greenlet==2.0.2
grpcio==1.53.0
h11==0.14.0
hmmlearn==0.2.8
huggingface-hub==0.12.0
HyperPyYAML==1.2.0
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
jellyfish==0.9.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.3
julius==0.2.7
kiwisolver==1.4.4
langcodes==3.3.0
librosa==0.9.2
lit==16.0.0
llvmlite==0.39.1
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
more-itertools==9.0.0
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.9
mutagen==1.46.0
networkx==2.8.8
nltk==3.8.1
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.3.0
openai-whisper==20230124
opentelemetry-api==1.14.0
opentelemetry-exporter-otlp-proto-http==1.14.0
opentelemetry-instrumentation==0.35b0
opentelemetry-instrumentation-aiohttp-client==0.35b0
opentelemetry-instrumentation-asgi==0.35b0
opentelemetry-proto==1.14.0
opentelemetry-sdk==1.14.0
opentelemetry-semantic-conventions==0.35b0
opentelemetry-util-http==0.35b0
optuna==3.1.1
packaging==23.0
pandas==1.5.3
pathspec==0.11.0
pathy==0.10.1
Pillow==9.4.0
pip-requirements-parser==32.0.1
pip-tools==6.12.2
platformdirs==3.2.0
pluggy==1.0.0
pooch==1.7.0
preshed==3.0.8
primePy==1.3
prometheus-client==0.16.0
protobuf==3.20.1
psutil==5.9.4
py==1.11.0
pyannote-whisper @ git+https://github.com/yinruiqing/pyannote-whisper.git@b9cd61164a371241f9aab387d7ae64e695ca2785
pyannote.audio==2.1.1
pyannote.core==4.5
pyannote.database==4.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==2.3
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodomex==3.17
pydantic==1.9.2
pydeck==0.8.0
pyDeprecate==0.3.2
Pygments==2.14.0
Pympler==1.0.1
pynvml==11.5.0
pyparsing==3.0.9
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pytest==7.1.1
python-dateutil==2.8.2
python-json-logger==2.0.6
python-multipart==0.0.5
pytorch-lightning==1.6.5
pytorch-metric-learning==1.7.3
pytube==12.1.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==25.0.0
regex==2022.3.15
requests==2.28.2
requests-oauthlib==1.3.1
resampy==0.4.2
rich==13.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
schema==0.7.5
scikit-learn==1.2.2
scipy==1.10.1
segtok==1.5.11
semantic-version==2.10.0
semver==2.13.0
sentencepiece==0.1.97
setuptools==61.2.0
shellingham==1.5.0.post1
simple-di==0.1.5
simple-term-menu==0.4.4
simplejson==3.19.1
singledispatchmethod==1.0
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sndfile==0.2.0
sniffio==1.3.0
sortedcontainers==2.4.0
SoundFile==0.10.3.post1
spacy==3.4.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
speechbrain==0.5.14
SQLAlchemy==2.0.9
srsly==2.4.5
starlette==0.26.1
streamlit==1.18.1
streamlit-player==0.1.5
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
thinc==8.1.7
threadpoolctl==3.1.0
tiktoken==0.3.1
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.1
torch-audiomentations==0.11.0
torch-pitch-shift==1.2.3
torchaudio==0.13.1
torchmetrics==0.11.4
torchvision==0.14.1
tornado==6.2
tqdm==4.64.0
transformers==4.24.0
triton==2.0.0
typer==0.4.2
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
uvicorn==0.20.0
validators==0.20.0
wasabi==0.10.1
watchdog==2.2.1
watchfiles==0.18.1
websocket-client==1.5.1
websockets==11.0.1
Werkzeug==2.2.3
wrapt==1.14.1
yake==0.4.8
yarl==1.8.2
yt-dlp==2023.3.4
zipp==3.14.0
